Add tests for `MissingTable`
