+++ causal-hub/tests/datasets/table/mod.rs
@@ -1,0 +1,2 @@
+mod categorical;
+mod missing;
